requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
openai>=1.68.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
SpeechRecognition==3.10.1
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.20
pytesseract==0.3.10
Pillow==10.4.0
tokenizers<=0.20.3,>=0.13.2
PyJWT==1.7.1 
requests>=2.31.0
types-requests>=2.31.0.20240125
psutil>=5.9.0 
requests[socks]==2.31.0